+++ src/library_glfw.js
@@ -653,7 +653,7 @@ var LibraryGLFW = {
             if (!GLFW.joys[joy]) {
               out('glfw joystick connected:',joy);
               GLFW.joys[joy] = {
-                id: allocate(intArrayFromString(gamepad.id)),
+                id: allocateUTF8(gamepad.id),
                 buttonsCount: gamepad.buttons.length,
                 axesCount: gamepad.axes.length,
                 buttons: _malloc(gamepad.buttons.length),
@@ -780,7 +780,7 @@ var LibraryGLFW = {
         };
         reader.readAsArrayBuffer(file);
 
-        var filename = allocate(intArrayFromString(path));
+        var filename = allocateUTF8(path);
         filenamesArray.push(filename);
         setValue(filenames + i*4, filename, 'i8*');
       }
@@ -1224,7 +1224,7 @@ var LibraryGLFW = {
   glfwGetVersionString__sig: 'i',
   glfwGetVersionString: function() {
     if (!GLFW.versionString) {
-      GLFW.versionString = allocate(intArrayFromString("3.2.1 JS WebGL Emscripten"));
+      GLFW.versionString = allocateUTF8("3.2.1 JS WebGL Emscripten");
     }
     return GLFW.versionString;
   },
@@ -1276,7 +1276,7 @@ var LibraryGLFW = {
   glfwGetMonitorName__sig: 'ii',
   glfwGetMonitorName: function(mon) {
     if (!GLFW.monitorString) {
-      GLFW.monitorString = allocate(intArrayFromString("HTML5 WebGL Canvas"));
+      GLFW.monitorString = allocateUTF8("HTML5 WebGL Canvas");
     }
     return GLFW.monitorString;
   },