+++ src/wasm2js.js
@@ -17,6 +17,7 @@ WebAssembly = {
   // Note that we do not use closure quoting (this['buffer'], etc.) on these
   // functions, as they are just meant for internal use. In other words, this is
   // not a fully general polyfill.
+  /** @constructor */
   Memory: function(opts) {
 #if USE_PTHREADS
     this.buffer = new SharedArrayBuffer(opts['initial'] * {{{ WASM_PAGE_SIZE }}});
@@ -60,6 +61,7 @@ WebAssembly = {
     // the main JS
   },
 
+  /** @constructor */
   Instance: function(module, info) {
     // TODO: use the module and info somehow - right now the wasm2js output is embedded in
     // the main JS